+++ test/src/expectEvent.web3.test.js
@@ -434,19 +434,32 @@ contract('expectEvent (web3 contracts) ', function ([deployer]) {
   });
 
   describe('not', function () {
+    describe('notEmitted', function () {
+      beforeEach(async function () {
+        this.receipt = await this.emitter.methods.emitArgumentless().send();
+      });
+
+      it('accepts not-emitted events', function () {
+        expectEvent.notEmitted(this.receipt, 'UnemittedEvent');
+      });
+
+      it('throws if an emitted event is requested', function () {
+        expect(() => expectEvent.notEmitted(this.receipt, 'Argumentless')).to.throw();
+      });
+    });
     describe('inTransaction', function () {
       context('with no arguments', function () {
         beforeEach(async function () {
           ({ transactionHash: this.txHash } = await this.emitter.methods.emitArgumentless().send());
         });
         it('accepts not emitted events', async function () {
-          await expectEvent.not.inTransaction(this.txHash, EventEmitter, 'WillNeverBeEmitted');
+          await expectEvent.notEmitted.inTransaction(this.txHash, EventEmitter, 'WillNeverBeEmitted');
         });
         it('throws when event does not exist in ABI', async function () {
-          await assertFailure(expectEvent.not.inTransaction(this.txHash, EventEmitter, 'Nonexistant'));
+          await assertFailure(expectEvent.notEmitted.inTransaction(this.txHash, EventEmitter, 'Nonexistant'));
         });
         it('throws when event its emitted', async function () {
-          await assertFailure(expectEvent.not.inTransaction(this.txHash, EventEmitter, 'Argumentless'));
+          await assertFailure(expectEvent.notEmitted.inTransaction(this.txHash, EventEmitter, 'Argumentless'));
         });
       });
       context('with arguments', function () {
@@ -455,10 +468,10 @@ contract('expectEvent (web3 contracts) ', function ([deployer]) {
           ({ transactionHash: this.txHash } = await this.emitter.methods.emitShortUint(this.value).send());
         });
         it('accepts not emitted events', async function () {
-          await expectEvent.not.inTransaction(this.txHash, EventEmitter, 'WillNeverBeEmitted');
+          await expectEvent.notEmitted.inTransaction(this.txHash, EventEmitter, 'WillNeverBeEmitted');
         });
         it('throws when event its emitted', async function () {
-          await assertFailure(expectEvent.not.inTransaction(this.txHash, EventEmitter, 'ShortUint'));
+          await assertFailure(expectEvent.notEmitted.inTransaction(this.txHash, EventEmitter, 'ShortUint'));
         });
       });
       context('with events emitted by an indirectly called contract', function () {
@@ -469,16 +482,18 @@ contract('expectEvent (web3 contracts) ', function ([deployer]) {
           ).send());
         });
         it('accepts not emitted events', async function () {
-          await expectEvent.not.inTransaction(this.txHash, EventEmitter, 'WillNeverBeEmitted');
+          await expectEvent.notEmitted.inTransaction(this.txHash, EventEmitter, 'WillNeverBeEmitted');
         });
         it('throws when event its emitted', async function () {
-          await assertFailure(expectEvent.not.inTransaction(this.txHash, IndirectEventEmitter, 'IndirectString'));
+          await assertFailure(
+            expectEvent.notEmitted.inTransaction(this.txHash, IndirectEventEmitter, 'IndirectString')
+          );
         });
       });
     });
     describe('inConstruction', function () {
       it('is unsupported', async function () {
-        await assertFailure(expectEvent.not.inConstruction(this.emitter, 'ShortUint'));
+        await assertFailure(expectEvent.notEmitted.inConstruction(this.emitter, 'ShortUint'));
       });
     });
   });